+++ account/cv_manager/views.py
@@ -124,7 +124,7 @@ def cv_panel_view(request):
     try:
         profile = user.userprofile
     except ObjectDoesNotExist:
-        return redirect('profile_create')
+        return redirect('create_profile')
 
     # Obtener CVs, experiencias, educaciones, idiomas, habilidades blandas y hobbies
     experiences = UserJobExperience.objects.filter(user=user)
@@ -176,28 +176,7 @@ def dashboard_view(request):
 
 
 
-# Vista para editar perfil usuario, incluye edición de User y UserProfile
-@login_required
-def edit_profile(request):
-    user = request.user
-    profile = user.userprofile
 
-    if request.method == 'POST':
-        user_form = UserForm(request.POST, instance=user)
-        profile_form = UserProfileForm(request.POST, request.FILES, instance=profile)
-
-        if user_form.is_valid() and profile_form.is_valid():
-            user_form.save()
-            profile_form.save()
-            return redirect('cv_panel')
-    else:
-        user_form = UserForm(instance=user)
-        profile_form = UserProfileForm(instance=profile)
-
-    return render(request, 'cv_manager/edit_profile.html', {
-        'user_form': user_form,
-        'profile_form': profile_form
-    })
 
 # Enlace a Experience_list
 @login_required
@@ -227,43 +206,6 @@ def delete_experience(request, pk):
         experience.delete()
         return redirect('experience_list')
     return render(request, 'cv_manager/experience_confirm_delete.html', {'experience': experience})
-
-
-# Vista para eliminar el perfil del usuario (solo perfil, no el usuario)
-@login_required
-def delete_userprofile(request):
-    profile = get_object_or_404(UserProfile, user=request.user)
-    
-    if request.method == 'POST':
-        profile.delete()
-        messages.success(request, "Tu perfil ha sido eliminado correctamente.")
-        return redirect('cv_panel')
-
-    return render(request, 'account/delete_userprofile_confirm.html', {'profile': profile})
-
-
-# Crear nuevo perfil (si no tiene)
-@login_required
-def create_profile_view(request):
-    user = request.user
-    try:
-        # Si ya tiene perfil, redirigir al panel
-        if user.userprofile:
-            return redirect('cv_panel')
-    except Exception:
-        pass
-
-    if request.method == 'POST':
-        form = UserProfileForm(request.POST, request.FILES)
-        if form.is_valid():
-            profile = form.save(commit=False)
-            profile.user = user
-            profile.save()
-            return redirect('cv_panel')
-    else:
-        form = CVProfileForm(user=request.user)
-
-    return render(request, 'account/create_profile.html', {'form': form})
 
 # Enlace a Education_list
 @login_required
