+++ account/templates/account/create_profile.html
@@ -3,7 +3,7 @@
 {% block title %}Crear Perfil{% endblock %}
 
 {% block content %}
-<h2>Crear perfil</h2>
+<h2>Crear perfil</h2><br>
 <form method="post" enctype="multipart/form-data">
   {% csrf_token %}
   {{ form.as_p }}
