+++ account/templates/account/delete_userprofile_confirm.html
@@ -9,6 +9,6 @@
 <form method="post">
   {% csrf_token %}
   <button type="submit" class="btn btn-danger">Eliminar mi perfil</button>
-  <a href="{% url 'cv_panel' %}" class="btn btn-secondary ms-2">Cancelar</a>
+  <a href="{% url 'create_profile' %}" class="btn btn-secondary ms-2">Cancelar</a>
 </form>
 {% endblock %}
