+++ account/views.py
@@ -13,6 +13,8 @@ from django.contrib.auth.forms import UserCreationForm
 from django.conf import settings
 from urllib.parse import urlparse, parse_qs
 
+from account.cv_manager.models import CVProfile
+from account.cv_manager.forms import UserProfileForm
 # External libraries
 import pdfkit
 import os
@@ -65,12 +67,71 @@ class UserLogoutView(LogoutView):
     next_page = reverse_lazy('login')
 
 # Vista para mostrar perfil del usuario, solo accesible con login
+# @login_required
+# def profile_view(request):
+#     user = request.user
+#     # Obtiene perfil del usuario o 404 si no existe
+#     perfil = get_object_or_404(UserProfile, user=user)
+#     # Obtiene CVs asociados al usuario
+#     cvs = CVProfile.objects.filter(user=user)
+#     return render(request, 'account/profile.html', {'perfil': perfil, 'cvs': cvs})
+
 @login_required
 def profile_view(request):
     user = request.user
-    # Obtiene perfil del usuario o 404 si no existe
-    perfil = get_object_or_404(UserProfile, user=user)
-    # Obtiene CVs asociados al usuario
+    # ✅ Si existe, devuelve el perfil; si no, devuelve None (sin error)
+    perfil = UserProfile.objects.filter(user=user).first()
+
     cvs = CVProfile.objects.filter(user=user)
+
     return render(request, 'account/profile.html', {'perfil': perfil, 'cvs': cvs})
 
+
+@login_required
+def edit_profile(request):
+    user = request.user
+    profile = user.userprofile
+    if request.method == 'POST':
+        user_form = UserForm(request.POST, instance=user)
+        profile_form = UserProfileForm(request.POST, request.FILES, instance=profile)
+        if user_form.is_valid() and profile_form.is_valid():
+            user_form.save()
+            profile_form.save()
+            return redirect('cv_panel')
+    else:
+        user_form = UserForm(instance=user)
+        profile_form = UserProfileForm(instance=profile)
+    return render(request, 'account/edit_profile.html', {
+        'user_form': user_form,
+        'profile_form': profile_form
+    })
+
+@login_required
+def delete_userprofile(request):
+    profile = get_object_or_404(UserProfile, user=request.user)
+    if request.method == 'POST':
+        profile.delete()
+        messages.success(request, "Tu perfil ha sido eliminado correctamente.")
+        return redirect('create_profile')
+    return render(request, 'account/delete_userprofile_confirm.html', {'profile': profile})
+
+@login_required
+def create_profile_view(request):
+    user = request.user
+    try:
+        if user.userprofile:
+            return redirect('cv_panel')
+    except Exception:
+        pass
+
+    if request.method == 'POST':
+        form = UserProfileForm(request.POST, request.FILES)
+        if form.is_valid():
+            profile = form.save(commit=False)
+            profile.user = user
+            profile.save()
+            return redirect('cv_panel')
+    else:
+        form = UserProfileForm()
+
+    return render(request, 'account/create_profile.html', {'form': form})
